+++ src/components/DocPages/Container.jsx
@@ -3,14 +3,8 @@ import * as styleVars from "../variable"
 
 export const Container = styled.section`
   display: flex;
+  max-width: 100%;
   align-items: center;
-  padding: 10rem 38rem 0 38rem;
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    display: block;
-    padding: 10%;
-
-  }
 `
 
 export default Container