+++ docs/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add Privacy Policy and Code of Conduct pages
+
 ## [1.2.3](https://github.com/hackmcgill/mchacks7/tree/1.2.3) - 2019-12-14
 
 ### Added